refactor: move migration to v2 migrations, modify `dtxp` column, and ignore migration for sqlite
